refactor: tidy up, move Profilefragment to fragment package
